+++ batch-console-api/src/test/java/io/github/pinpols/batch/console/domain/rbac/web/ConsoleMetaControllerTest.java
@@ -9,7 +9,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import io.github.pinpols.batch.common.dto.ResponseMeta;
 import io.github.pinpols.batch.common.time.BatchDateTimeSupport;
-import io.github.pinpols.batch.console.domain.rbac.service.ConsoleMetaQueryService;
+import io.github.pinpols.batch.console.application.rbac.ConsoleMetaQueryService;
 import io.github.pinpols.batch.console.domain.rbac.web.response.ConsoleMetaEnumItem;
 import io.github.pinpols.batch.console.domain.rbac.web.response.ConsoleMetaOption;
 import io.github.pinpols.batch.console.service.ConsoleResponseFactory;